Get Count of Clinical Trials from ClinicalTrials.gov

This function retrieves the count of clinical trials from ClinicalTrials.gov based on specified parameters.

Usage

ctg_count(
  condition = NULL,
  location = NULL,
  title = NULL,
  intervention = NULL,
  status = NULL
)

Arguments

condition

A character string specifying the condition being studied (default: NULL).

location

A character string specifying the location of the trials (default: NULL).

title

A character string specifying keywords in the study title (default: NULL).

intervention

A character string specifying the type of intervention (default: NULL).

status

A character vector specifying the recruitment status of the trials. Allowed values are: Valid values include:

ACTIVE_NOT_RECRUITING - Studies that are actively conducting but not recruiting participants.

COMPLETED - Studies that have completed all phases.

ENROLLING_BY_INVITATION - Studies that are enrolling participants by invitation only.

NOT_YET_RECRUITING - Studies that have not yet started recruiting.

RECRUITING - Studies that are actively recruiting participants.

SUSPENDED - Studies that are temporarily halted.

TERMINATED - Studies that have been terminated before completion.

WITHDRAWN - Studies that have been withdrawn before enrollment.

AVAILABLE - Studies that are available.

NO_LONGER_AVAILABLE - Studies that are no longer available.

TEMPORARILY_NOT_AVAILABLE - Studies that are temporarily not available.

APPROVED_FOR_MARKETING - Studies that have been approved for marketing.

WITHHELD - Studies that have data withheld.

UNKNOWN - Studies with an unknown status.

Default is NULL.

Value

A number representing the total count of clinical trials matching the specified parameters.

Examples

ctg_count(
  condition = "Cancer",
  location = "India",
  title = NULL,
  intervention = "Drug",
  status = "RECRUITING"
)
#> [1] 101