Query ClinicalTrials.gov API

This function sends a query to the ClinicalTrials.gov API and returns the results as a tibble. Users can specify various parameters to filter the results, and if a parameter is not provided, it will be omitted from the query.

Usage

query_clinical_trials(
  condition = "Glioblastoma",
  location = "India",
  title = NULL,
  intervention = "Drug",
  status = NULL,
  page_size = 20
)

Arguments

condition

A character string specifying the medical condition to search for. This will filter the results to studies related to the given condition.

location

A character string specifying the location (e.g., city or country) to search in. This will filter the results to studies conducted in the specified location.

title

A character string specifying keywords to search for in study title. This will filter the results to studies with title that include the specified keywords.

intervention

A character string specifying the intervention or treatment to search for. This will filter the results to studies involving the specified intervention.

status

A character vector specifying the overall status of the studies. Valid values include:

ACTIVE_NOT_RECRUITING - Studies that are actively conducting but not recruiting participants.

COMPLETED - Studies that have completed all phases.

ENROLLING_BY_INVITATION - Studies that are enrolling participants by invitation only.

NOT_YET_RECRUITING - Studies that have not yet started recruiting.

RECRUITING - Studies that are actively recruiting participants.

SUSPENDED - Studies that are temporarily halted.

TERMINATED - Studies that have been terminated before completion.

WITHDRAWN - Studies that have been withdrawn before enrollment.

AVAILABLE - Studies that are available.

NO_LONGER_AVAILABLE - Studies that are no longer available.

TEMPORARILY_NOT_AVAILABLE - Studies that are temporarily not available.

APPROVED_FOR_MARKETING - Studies that have been approved for marketing.

WITHHELD - Studies that have data withheld.

UNKNOWN - Studies with an unknown status.

page_size

An integer specifying the number of results per page. The default value is 20. The maximum allowed value is 1,000. If a value greater than 1,000 is specified, it will be coerced to 1,000. If not specified, the default value will be used.

Value

A tibble containing the query results. Each row represents a study, and the columns correspond to the study details returned by the API.

Details

This function can return up to 1,000 results.

The function constructs a query to the ClinicalTrials.gov API using the provided parameters. It supports filtering by condition, location, title keywords, intervention, and overall status. The function handles the API response, checks for errors, and parses the results into a tibble.

Examples

# Query for studies related to "diabetes" in "Kolkata" with the status "RECRUITING"
query_clinical_trials(condition = "diabetes", location = "Kolkata",
                                 status = "RECRUITING")
#> The Query matches 9 trial records in the ClinicalTrials.gov records.
#> Your query returned 9 trial records.
#> # A tibble: 9 × 30
#>   `NCT Number` `Study Title`  `Study URL` Acronym `Study Status` `Brief Summary`
#>   <chr>        <chr>          <chr>       <chr>   <chr>          <chr>          
#> 1 NCT04596631  A Research St… https://cl… PIONEE… RECRUITING     "This study co…
#> 2 NCT06065540  A Research St… https://cl… REIMAG… RECRUITING     "The study wil…
#> 3 NCT05929079  A Study of Re… https://cl… TRIUMP… RECRUITING     "The purpose o…
#> 4 NCT06221969  A Research St… https://cl… NA      RECRUITING     "This study wi…
#> 5 NCT06370715  A Study of LY… https://cl… NA      RECRUITING     "The purpose o…
#> 6 NCT05254002  A Study to Le… https://cl… CONFID… RECRUITING     "Finerenone wo…
#> 7 NCT05705271  A Study to Le… https://cl… NA      RECRUITING     "Researchers a…
#> 8 NCT06269107  A Research St… https://cl… COMBIN… RECRUITING     "This study wi…
#> 9 NCT05929066  A Study of Re… https://cl… TRIUMP… RECRUITING     "The purpose o…
#> # ℹ 24 more variables: `Study Results` <chr>, Conditions <chr>,
#> #   Interventions <chr>, `Primary Outcome Measures` <chr>,
#> #   `Secondary Outcome Measures` <chr>, `Other Outcome Measures` <lgl>,
#> #   Sponsor <chr>, Collaborators <chr>, Sex <chr>, Age <chr>, Phases <chr>,
#> #   Enrollment <dbl>, `Funder Type` <chr>, `Study Type` <chr>,
#> #   `Study Design` <chr>, `Other IDs` <chr>, `Start Date` <date>,
#> #   `Primary Completion Date` <chr>, `Completion Date` <chr>, …

# Query for studies with "vaccine" in the title and the status "COMPLETED"
query_clinical_trials(title = "vaccine", status = "COMPLETED", page_size = 50)
#> The Query matches 0 trial records in the ClinicalTrials.gov records.
#> Your query returned 0 trial records.
#> # A tibble: 0 × 30
#> # ℹ 30 variables: NCT Number <chr>, Study Title <chr>, Study URL <chr>,
#> #   Acronym <chr>, Study Status <chr>, Brief Summary <chr>,
#> #   Study Results <chr>, Conditions <chr>, Interventions <chr>,
#> #   Primary Outcome Measures <chr>, Secondary Outcome Measures <chr>,
#> #   Other Outcome Measures <chr>, Sponsor <chr>, Collaborators <chr>,
#> #   Sex <chr>, Age <chr>, Phases <chr>, Enrollment <chr>, Funder Type <chr>,
#> #   Study Type <chr>, Study Design <chr>, Other IDs <chr>, Start Date <chr>, …